+++ app/src/main/java/com/example/ap2_ex3/converters/DisplayedUserConverter.java
@@ -11,7 +11,6 @@ public class DisplayedUserConverter {
     @TypeConverter
     public DisplayedUser StringToDisplayedUser(String value) {
         List<String> userList = Arrays.asList(value.split(","));
-
         return new DisplayedUser(userList.get(0), userList.get(1), userList.get(2));
     }
